Installing unit test gem and setting up folders for testing
