+++ Gemfile
@@ -3,5 +3,6 @@ source "https://rubygems.org"
 
 gem 'spreadsheet'
 gem 'locale'
+gem 'test-unit'
 gem 'i18n'
 gem 'qif'
